+++ docs/news.rst
@@ -1,6 +1,11 @@
 Release Notes
 =============
 
+**UNRELEASED**
+
+- Added the ``wheel info`` subcommand to display metadata about wheel files without
+  unpacking them (`#639 <https://github.com/pypa/wheel/issues/639>`_)
+
 **0.46.3 (2026-01-22)**
 
 - Fixed ``ImportError: cannot import name '_setuptools_logging' from 'wheel'`` when
@@ -9,8 +14,6 @@ Release Notes
 
 **0.46.2 (2026-01-22)**
 
-- Added the ``wheel info`` subcommand to display metadata about wheel files without
-  unpacking them (`#639 <https://github.com/pypa/wheel/issues/639>`_)
 - Restored the ``bdist_wheel`` command for compatibility with ``setuptools`` older than
   v70.1
 - Importing ``wheel.bdist_wheel`` now emits a ``FutureWarning`` instead of a